Move all models into their own file.
