+++ starter_kits/Swift/Halite-III/Halite-III/hlt/Game.swift
@@ -1,0 +1,25 @@
+//
+//  Game.swift
+//  Halite-III
+//
+//  Created by Chris Downie on 10/24/18.
+//  Copyright © 2018 Chris Downie. All rights reserved.
+//
+
+import Foundation
+
+class Game {
+    var turnNumber = 0
+    
+    func ready(botName: String) {
+        
+    }
+    
+    func updateFrame() {
+        
+    }
+    
+    func endTurn(commands: [Command]) {
+        
+    }
+}
