+++ starter_kits/Swift/Halite-III/Halite-III/hlt/Log.swift
@@ -1,3 +1,10 @@
+//
+//  Log.swift
+//  Halite-III
+//
+//  Created by Chris Downie on 10/24/18.
+//  Copyright © 2018 Chris Downie. All rights reserved.
+//
 import Foundation
 
 class Log {
