+++ starter_kits/Swift/Halite-III/Halite-III/main.swift
@@ -3,20 +3,5 @@ import Foundation
 // Load constants
 _ = Constant.shared
 
-class Game {
-    var turnNumber = 0
-    
-    func ready(botName: String) {
-        
-    }
-    
-    func updateFrame() {
-        
-    }
-    
-    func endTurn(commands: [Command]) {
-        
-    }
-}
 
 let game = Game()
